sync heading syntax with meetup.rst, remove question marks
Standardizing on the syntax mentioned in http://www.sphinx-doc.org/en/stable/rest.html#sections
and comes from https://devguide.python.org/documenting/#sections

= for sections
- for subsections
^ for subsubsections

It was already implemented in several texts here. I'm not really
opinionated about it, open to discussion and all, but we should be
at least consistent in this.